Do not validate email addresses
This package is dumb, in that the value object class does not try and validate the email address strings passed into the constructor. The `FILTER_VALIDATE_EMAIL` misses a lot of valid email addresses.

Better to use https://github.com/egulias/EmailValidator or similar for validation.

The `$address` is no longer stored within the class. Closes #1.